ch: refactor redis_stream to include polling
 - refactor redis_stream to include polling

[Finishes #172825140]
